feat: change cat service

+++ service/src/main/java/ru/kramskoi/service/CatServiceImpl.java
@@ -6,17 +6,23 @@ import ru.kramskoi.breeds.Breed;
 import ru.kramskoi.colors.Color;
 import ru.kramskoi.dto.CatDTO;
 import ru.kramskoi.entity.Cat;
+import ru.kramskoi.entity.Person;
+import ru.kramskoi.exception.CatNotFound;
 import ru.kramskoi.mapper.CatMapper;
 import ru.kramskoi.repository.CatRepository;
+import ru.kramskoi.repository.PersonRepository;
 
+import java.security.Principal;
 import java.util.List;
 
 @Service
 public class CatServiceImpl implements CatService {
     private final CatRepository catRepository;
+    private final PersonRepository personRepository;
 
-    public CatServiceImpl(CatRepository catRepository) {
+    public CatServiceImpl(CatRepository catRepository, PersonRepository personRepository) {
         this.catRepository = catRepository;
+        this.personRepository = personRepository;
     }
 
     @Override
@@ -27,36 +33,44 @@ public class CatServiceImpl implements CatService {
 
     @Override
     @Transactional(readOnly = true)
-    public CatDTO findCatByID(Long id) {
+    public CatDTO findCatByID(Long id, Principal principal) {
         Cat cat = catRepository.getCatById(id);
         if (cat == null) {
-            throw new IllegalArgumentException();
+            throw new CatNotFound();
         }
+
+        String username = principal.getName();
+        Person person = personRepository.findByUsername(username);
+        List<Cat> ownersCats = person.getOwner().getCats();
+        if (ownersCats.stream().findFirst().filter(c -> c.getId() == cat.getId()).isEmpty()) {
+            throw new CatNotFound();
+        }
+
         return CatMapper.fromCatToDTO(cat);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public List<CatDTO> findCatsByColor(Color color) {
-        List<CatDTO> catDTOList = catRepository.getCatsByColor(color)
+    public List<CatDTO> findCatsByColor(Color color, Principal principal) {
+        List<CatDTO> catDTOList = new java.util.ArrayList<>(catRepository.getCatsByColor(color)
                 .stream()
                 .map(CatMapper::fromCatToDTO)
-                .toList();
+                .toList());
         if (catDTOList.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new CatNotFound();
         }
         return catDTOList;
     }
 
     @Override
     @Transactional(readOnly = true)
-    public List<CatDTO> findCatsByBreed(Breed breed) {
+    public List<CatDTO> findCatsByBreed(Breed breed, Principal principal) {
         List<CatDTO> catDTOList = catRepository.getCatsByBreed(breed)
                 .stream()
                 .map(CatMapper::fromCatToDTO)
                 .toList();
         if (catDTOList.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new CatNotFound();
         }
         return catDTOList;
     }
@@ -64,42 +78,73 @@ public class CatServiceImpl implements CatService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<CatDTO> getCatsByColorOrBreed(Color color, Breed breed) {
+    public List<CatDTO> getCatsByColorOrBreed(Color color, Breed breed, Principal principal) {
         if (breed == null) {
-            return findCatsByColor(color);
+            return findCatsByColor(color, principal);
         } else if (color == null) {
-            return findCatsByBreed(breed);
+            return findCatsByBreed(breed, principal);
         }
 
         List<CatDTO> catDTOList = catRepository.getCatsByColorOrBreed(color, breed)
                 .stream()
                 .map(CatMapper::fromCatToDTO)
                 .toList();
+
+        List<Long> ownersCats = personRepository.findByUsername(principal.getName())
+                .getOwner()
+                .getCats()
+                .stream()
+                .map(Cat::getId)
+                .toList();
+
+        catDTOList.removeIf(catDTO -> !ownersCats.contains(catDTO.getId()));
         if (catDTOList.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new CatNotFound();
         }
         return catDTOList;
     }
 
     @Override
     @Transactional(readOnly = true)
-    public List<CatDTO> getFriendsById(Long id) {
+    public List<CatDTO> getFriendsById(Long id, Principal principal) {
         List<CatDTO> catDTOList = catRepository.getCatById(id)
                 .getFriends()
                 .stream()
                 .map(CatMapper::fromCatToDTO)
                 .toList();
         if (catDTOList.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new CatNotFound();
         }
+
+        List<Long> ownersCats = personRepository.findByUsername(principal.getName())
+                .getOwner()
+                .getCats()
+                .stream()
+                .map(Cat::getId)
+                .toList();
+
+        catDTOList.removeIf(catDTO -> !ownersCats.contains(catDTO.getId()));
+        if (catDTOList.isEmpty()) {
+            throw new CatNotFound();
+        }
+
         return catDTOList;
     }
 
     @Override
     @Transactional
-    public void addFriend(Long catId, Long friendId) {
+    public void addFriend(Long catId, Long friendId, Principal principal) {
         Cat cat = catRepository.getCatById(catId);
         Cat friend = catRepository.getCatById(friendId);
+        List<Long> ownersCats = personRepository.findByUsername(principal.getName())
+                .getOwner()
+                .getCats()
+                .stream()
+                .map(Cat::getId)
+                .toList();
+        if (!ownersCats.contains(cat.getId())) {
+            throw new CatNotFound();
+        }
         cat.addFriend(friend);
         friend.addFriend(cat);
         catRepository.save(cat);
@@ -108,21 +153,41 @@ public class CatServiceImpl implements CatService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<CatDTO> getAllCatsByOwnerId(Long id) {
+    public List<CatDTO> getAllCatsByOwnerId(Long id, Principal principal) {
         List<CatDTO> catDTOList = catRepository.getCatsByOwnerId(id)
                 .stream()
                 .map(CatMapper::fromCatToDTO)
                 .toList();
+
+        List<Long> ownersCats = personRepository.findByUsername(principal.getName())
+                .getOwner()
+                .getCats()
+                .stream()
+                .map(Cat::getId)
+                .toList();
+
+        catDTOList.removeIf(catDTO -> !ownersCats.contains(catDTO.getId()));
         if (catDTOList.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new CatNotFound();
         }
+
         return catDTOList;
     }
 
     @Override
     @Transactional
-    public void updateCat(Cat cat) {
+    public void updateCat(Cat cat, Principal principal) {
         Cat prevCat = catRepository.getCatById(cat.getId());
+        List<Long> ownersCats = personRepository.findByUsername(principal.getName())
+                .getOwner()
+                .getCats()
+                .stream()
+                .map(Cat::getId)
+                .toList();
+        if (!ownersCats.contains(cat.getId())) {
+            throw new CatNotFound();
+        }
+
         prevCat.setName(cat.getName());
         prevCat.setBirthday(cat.getBirthday());
         prevCat.setBreed(cat.getBreed());
@@ -132,7 +197,16 @@ public class CatServiceImpl implements CatService {
 
     @Override
     @Transactional
-    public void deleteCat(Long id) {
+    public void deleteCat(Long id, Principal principal) {
+        List<Long> ownersCats = personRepository.findByUsername(principal.getName())
+                .getOwner()
+                .getCats()
+                .stream()
+                .map(Cat::getId)
+                .toList();
+        if (!ownersCats.contains(id)) {
+            throw new CatNotFound();
+        }
         catRepository.deleteById(id);
     }
 }
